pyproject.toml
[tool.poetry]
name = "nonebot-plugin-pixivbot-kook"
version = "1.1.0-alpha2"
description = "Nonebot Plugin Pixivbot (KOOK)"
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/ssttkkl/nonebot-plugin-pixivbot-kook"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-beta.4"
nonebot-adapter-kaiheila = "^0.0.2"
nonebot-plugin-pixivbot = "1.1.0-alpha.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[build-system]
requires = ["setuptools~=61.0"]
build-backend = "setuptools.build_meta"